additions/sctp-client/requirements.txt
pysctp==0.7.2
